refactor(express): use new llamaindex for sharing chat logic between ts frameworks
