+++ helpers/typescript.ts
@@ -104,6 +104,17 @@ export const installTSTemplate = async ({
       : path.join("src", "controllers");
   const enginePath = path.join(root, relativeEngineDestPath, "engine");
 
+  // copy llamaindex folder for TS templates
+  console.log(
+    { framework },
+    relativeEngineDestPath,
+    path.join(compPath, "shared", "llamaindex"),
+  );
+  await copy("**", path.join(root, relativeEngineDestPath, "llamaindex"), {
+    parents: true,
+    cwd: path.join(compPath, "shared", "llamaindex"),
+  });
+
   // copy vector db component
   if (vectorDb === "llamacloud") {
     console.log(
