+++ templates/components/shared/llamaindex/embeddings.ts
