+++ templates/components/shared/llamaindex/events.ts
