+++ templates/components/shared/llamaindex/stream.ts
