+++ templates/types/streaming/express/src/controllers/chat-upload.controller.ts
@@ -1,12 +1,12 @@
 import { Request, Response } from "express";
-import { readAndSplitDocument } from "./embeddings";
+import { uploadDocument } from "./llamaindex/embeddings";
 
-export const chatEmbed = async (req: Request, res: Response) => {
+export const chatUpload = async (req: Request, res: Response) => {
   const { base64 }: { base64: string } = req.body;
   if (!base64) {
     return res.status(400).json({
       error: "base64 is required in the request body",
     });
   }
-  return res.status(200).json(await readAndSplitDocument(base64));
+  return res.status(200).json(await uploadDocument(base64));
 };
