+++ templates/types/streaming/express/src/routes/chat.route.ts
@@ -1,7 +1,7 @@
 import express, { Router } from "express";
 import { chatConfig } from "../controllers/chat-config.controller";
-import { chatEmbed } from "../controllers/chat-embed.controller";
 import { chatRequest } from "../controllers/chat-request.controller";
+import { chatUpload } from "../controllers/chat-upload.controller";
 import { chat } from "../controllers/chat.controller";
 import { initSettings } from "../controllers/engine/settings";
 
@@ -11,6 +11,6 @@ initSettings();
 llmRouter.route("/").post(chat);
 llmRouter.route("/request").post(chatRequest);
 llmRouter.route("/config").get(chatConfig);
-llmRouter.route("/embed").post(chatEmbed);
+llmRouter.route("/upload").post(chatUpload);
 
 export default llmRouter;
